// packages/typespec-test/test/translator/generated/typespec-ts/src/models/models.ts
export interface GetLanguagesResult {
  /** Languages that support translate API. */
  translation: Record<string, TranslationLanguage>;
  /** Languages that support transliteration API. */
  transliteration: Record<string, TransliterationLanguage>;
  /** Languages that support dictionary API. */
  dictionary: Record<string, SourceDictionaryLanguage>;
}

/**
 * The value of the translation property is a dictionary of (key, value) pairs. Each key is a BCP 47 language tag.
 * A key identifies a language for which text can be translated to or translated from.
 */
export interface TranslationLanguage {
  /** Display name of the language in the locale requested via Accept-Language header. */
  name: string;
  /** Display name of the language in the locale native for this language. */
  nativeName: string;
  /** Directionality, which is rtl for right-to-left languages or ltr for left-to-right languages. */
  dir: string;
}

/**
 * The value of the transliteration property is a dictionary of (key, value) pairs.
 * Each key is a BCP 47 language tag. A key identifies a language for which text can be converted from one script
 * to another script.
 */
export interface TransliterationLanguage {
  /** Display name of the language in the locale requested via Accept-Language header. */
  name: string;
  /** Display name of the language in the locale native for this language. */
  nativeName: string;
  /** List of scripts to convert from. */
  scripts: TransliterableScript[];
}
// ...
export interface TransliterableScript extends CommonScriptModel {
  /** List of scripts available to convert text to. */
  toScripts: CommonScriptModel[];
}

/** Properties ot the source dictionary language */
export interface SourceDictionaryLanguage {
  /** Display name of the language in the locale requested via Accept-Language header. */
  name: string;
  /** Display name of the language in the locale native for this language. */
  nativeName: string;
  /** Directionality, which is rtl for right-to-left languages or ltr for left-to-right languages. */
  dir: string;
  /** List of languages with alterative translations and examples for the query expressed in the source language. */
  translations: TargetDictionaryLanguage[];
}

/** Properties of the target dictionary language */
export interface TargetDictionaryLanguage {
  /** Display name of the language in the locale requested via Accept-Language header. */
  name: string;
  /** Display name of the language in the locale native for this language. */
  nativeName: string;
  /** Directionality, which is rtl for right-to-left languages or ltr for left-to-right languages. */
  dir: string;
  /** Language code identifying the target language. */
  code: string;
}
// ...
export interface InputTextElement {
  /** Text to translate. */
  text: string;
}

/** Element containing the translated text */
export interface TranslatedTextElement {
  /** The detectedLanguage property is only present in the result object when language auto-detection is requested. */
  detectedLanguage?: DetectedLanguage;
  /**
   * An array of translation results. The size of the array matches the number of target
   * languages specified through the to query parameter.
   */
  translations: Translation[];
  /**
   * Input text in the default script of the source language. sourceText property is present only when
   * the input is expressed in a script that's not the usual script for the language. For example,
   * if the input were Arabic written in Latin script, then sourceText.text would be the same Arabic text
   * converted into Arab script.
   */
  sourceText: SourceText;
}

/** An object describing the detected language. */
export interface DetectedLanguage {
  /** A string representing the code of the detected language. */
  language: string;
  /**
   * A float value indicating the confidence in the result.
   * The score is between zero and one and a low score indicates a low confidence.
   */
  score: number;
}

/** Translation result */
export interface Translation {
  /** A string representing the language code of the target language. */
  to: string;
  /** A string giving the translated text. */
  text: string;
  /** An object giving the translated text in the script specified by the toScript parameter. */
  transliteration: Transliteration;
  /** Alignment information. */
  alignment: Alignment;
  /** Sentence boundaries in the input and output texts. */
  sentLen: SentenceLength;
}

/** An object giving the translated text in the script specified by the toScript parameter. */
export interface Transliteration {
  /** A string specifying the target script. */
  script: string;
  /** A string giving the translated text in the target script. */
  text: string;
}

/** Alignment information object. */
export interface Alignment {
  /**
   * Maps input text to translated text. The alignment information is only provided when the request
   * parameter includeAlignment is true. Alignment is returned as a string value of the following
   * format: [[SourceTextStartIndex]:[SourceTextEndIndex]–[TgtTextStartIndex]:[TgtTextEndIndex]].
   * The colon separates start and end index, the dash separates the languages, and space separates the words.
   * One word may align with zero, one, or multiple words in the other language, and the aligned words may
   * be non-contiguous. When no alignment information is available, the alignment element will be empty.
   */
  proj: string;
}

/** An object returning sentence boundaries in the input and output texts. */
export interface SentenceLength {
  /**
   * An integer array representing the lengths of the sentences in the input text.
   * The length of the array is the number of sentences, and the values are the length of each sentence.
   */
  srcSentLen: number[];
  /**
   * An integer array representing the lengths of the sentences in the translated text.
   * The length of the array is the number of sentences, and the values are the length of each sentence.
   */
  transSentLen: number[];
}

/** Input text in the default script of the source language. */
export interface SourceText {
  /** Input text in the default script of the source language. */
  text: string;
}

/** Transliterated text element. */
export interface TransliteratedText {
  /** A string which is the result of converting the input string to the output script. */
  text: string;
  /** A string specifying the script used in the output. */
  script: string;
}

/** Detected language */
export interface DetectedLanguageElement {
  /** Code of the detected language. */
  language: string;
  /**
   * A float value indicating the confidence in the result. The score is between zero and one
   * and a low score indicates a low confidence.
   */
  score: number;
  /** A boolean value which is true if the detected language is one of the languages supported for text translation. */
  isTranslationSupported: boolean;
  /** A boolean value which is true if the detected language is one of the languages supported for transliteration. */
  isTransliterationSupported: boolean;
}

/** Detected Language with alternatives */
export interface DetectedLanguageWithAlternatives
  extends DetectedLanguageElement {
  /** An array of other possible languages. */
  alternatives: DetectedLanguage[];
}

/** Elemented containing break sentence result. */
export interface BreakSentenceElement {
  /** The detectedLanguage property is only present in the result object when language auto-detection is requested. */
  detectedLanguage?: DetectedLanguage;
  /** Sentence boundaries in the input and output texts. */
  sentLen: SentenceLength;
}

/** Dictionary Lookup Element */
export interface DictionaryLookupElement {
  /**
   * A string giving the normalized form of the source term.
   * For example, if the request is "JOHN", the normalized form will be "john".
   * The content of this field becomes the input to lookup examples.
   */
  normalizedSource: string;
  /**
   * A string giving the source term in a form best suited for end-user display.
   * For example, if the input is "JOHN", the display form will reflect the usual
   * spelling of the name: "John".
   */
  displaySource: string;
  /** A list of translations for the source term. */
  translations: DictionaryTranslation[];
}

/** Translation source term. */
export interface DictionaryTranslation {
  /**
   * A string giving the normalized form of this term in the target language.
   * This value should be used as input to lookup examples.
   */
  normalizedTarget: string;
  /**
   * A string giving the term in the target language and in a form best suited
   * for end-user display. Generally, this will only differ from the normalizedTarget
   * in terms of capitalization. For example, a proper noun like "Juan" will have
   * normalizedTarget = "juan" and displayTarget = "Juan".
   */
  displayTarget: string;
  /** A string associating this term with a part-of-speech tag. */
  posTag: string;
  /**
   * A value between 0.0 and 1.0 which represents the "confidence"
   * (or perhaps more accurately, "probability in the training data") of that translation pair.
   * The sum of confidence scores for one source word may or may not sum to 1.0.
   */
  confidence: number;
  /**
   * A string giving the word to display as a prefix of the translation. Currently,
   * this is the gendered determiner of nouns, in languages that have gendered determiners.
   * For example, the prefix of the Spanish word "mosca" is "la", since "mosca" is a feminine noun in Spanish.
   * This is only dependent on the translation, and not on the source.
   * If there is no prefix, it will be the empty string.
   */
  prefixWord: string;
  /**
   * A list of "back translations" of the target. For example, source words that the target can translate to.
   * The list is guaranteed to contain the source word that was requested (e.g., if the source word being
   * looked up is "fly", then it is guaranteed that "fly" will be in the backTranslations list).
   * However, it is not guaranteed to be in the first position, and often will not be.
   */
  backTranslations: BackTranslation[];
}
// ...
export interface DictionaryExampleTextElement extends InputTextElement {
  /**
   * A string specifying the translated text previously returned by the Dictionary lookup operation.
   * This should be the value from the normalizedTarget field in the translations list of the Dictionary
   * lookup response. The service will return examples for the specific source-target word-pair.
   */
  translation: string;
}

/** Dictionary Example element */
export interface DictionaryExampleElement {
  /**
   * A string giving the normalized form of the source term. Generally, this should be identical
   * to the value of the Text field at the matching list index in the body of the request.
   */
  normalizedSource: string;
  /**
   * A string giving the normalized form of the target term. Generally, this should be identical
   * to the value of the Translation field at the matching list index in the body of the request.
   */
  normalizedTarget: string;
  /** A list of examples for the (source term, target term) pair. */
  examples: Example[];
}

/** Dictionary Example */
export interface Example {
  /**
   * The string to concatenate before the value of sourceTerm to form a complete example.
   * Do not add a space character, since it is already there when it should be.
   * This value may be an empty string.
   */
  sourcePrefix: string;
  /**
   * A string equal to the actual term looked up. The string is added with sourcePrefix
   * and sourceSuffix to form the complete example. Its value is separated so it can be
   * marked in a user interface, e.g., by bolding it.
   */
  sourceTerm: string;
  /**
   * The string to concatenate after the value of sourceTerm to form a complete example.
   * Do not add a space character, since it is already there when it should be.
   * This value may be an empty string.
   */
  sourceSuffix: string;
  /** A string similar to sourcePrefix but for the target. */
  targetPrefix: string;
  /** A string similar to sourceTerm but for the target. */
  targetTerm: string;
  /** A string similar to sourceSuffix but for the target. */
  targetSuffix: string;
}